requirements.txt
Pillow==10.0.0
torch==2.0.1
numpy==1.24.3
torchvision==0.15.2
onnxruntime==1.15.1
requests==2.31.0